django/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
--index-url https://artifactory.klarna.net/artifactory/api/pypi/r-pypi-pypi-org/simple
--extra-index-url https://artifactory.klarna.net/artifactory/api/pypi/v-pypi-production/simple/

aiofiles==23.2.1
    # via python-schema-registry-client
anyio==4.3.0
    # via httpx
asgiref==3.7.2
    # via django
async-timeout==4.0.3
    # via redis
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.0.1
    # via django (pyproject.toml)
certifi==2023.11.17
    # via
    #   django-healthchecks
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
confluent-kafka==2.3.0
    # via django (pyproject.toml)
cryptography==41.0.5
    # via pyjwt
django==4.2.7
    # via
    #   django (pyproject.toml)
    #   django-axes
    #   django-filter
    #   django-healthchecks
    #   django-storages
    #   djangorestframework
    #   drf-jwt
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-axes==6.1.1
    # via django (pyproject.toml)
django-behaviors==0.5.1
    # via django (pyproject.toml)
django-environ==0.11.2
    # via django (pyproject.toml)
django-filter==23.4
    # via django (pyproject.toml)
django-healthchecks==1.5.0
    # via django (pyproject.toml)
django-ipware==5.0.2
    # via django (pyproject.toml)
django-split-settings==1.2.0
    # via django (pyproject.toml)
django-storages==1.14.2
    # via django (pyproject.toml)
djangorestframework==3.14.0
    # via
    #   django (pyproject.toml)
    #   drf-jwt
    #   drf-spectacular
djangorestframework-camel-case==1.4.2
    # via django (pyproject.toml)
drf-jwt==1.19.2
    # via django (pyproject.toml)
drf-spectacular[sidecar]==0.26.5
    # via django (pyproject.toml)
drf-spectacular-sidecar==2023.10.1
    # via drf-spectacular
exceptiongroup==1.2.0
    # via anyio
fastavro==1.9.4
    # via python-schema-registry-client
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.26.0
    # via python-schema-registry-client
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.20.0
    # via
    #   drf-spectacular
    #   python-schema-registry-client
jsonschema-specifications==2023.11.1
    # via jsonschema
pillow==10.1.0
    # via django (pyproject.toml)
psycopg2-binary==2.9.9
    # via django (pyproject.toml)
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via
    #   drf-jwt
    #   pyjwt
python-schema-registry-client==2.5.6
    # via django (pyproject.toml)
pytz==2023.3.post1
    # via djangorestframework
pyyaml==6.0.1
    # via drf-spectacular
redis==5.0.1
    # via django (pyproject.toml)
referencing==0.31.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via django-healthchecks
rpds-py==0.13.1
    # via
    #   jsonschema
    #   referencing
sentry-sdk==1.35.0
    # via django (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlparse==0.4.4
    # via django
tenacity==8.2.3
    # via django (pyproject.toml)
typing-extensions==4.9.0
    # via
    #   anyio
    #   asgiref
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.1.0
    # via
    #   requests
    #   sentry-sdk
whitenoise==6.6.0
    # via django (pyproject.toml)

# The following packages are considered to be unsafe in a requirements file:
# setuptools
